Zombies are walking, but broke it when working on getting them to walk towards and then stop at the barricade

+++ ZombieDefenceShooter/Assets/Scripts/Components/GameProperties.cs
@@ -11,7 +11,7 @@ namespace ZDS_DOTS
         public float2 gameObjPos;
         public Entity spawnPointPrefab;
         public Entity zombiePrefab;
-        public uint spawnRate;
+        public float spawnRate;
     }
 
     public struct SpawnTimer : IComponentData
